// poc-aw-ngrx/ClientApp/src/app/models/achAccount.ts
export class AchAccount {
  taxId: number;
  routingNumber: number;
  accountName: string;
  name: string;
  accountType: number;
  type: string;
  accountNumber: string;
  status: string;
  bankName: string;
  institution: string;
  prenoteDate: string;
  sequenceId: number;
  disabledYN: number;
  updated: string;
}